+++ app/src/main/java/io/github/quinnandrews/scheduler/modules/administration/schedules/core/domain/repositories/LocationRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 
 @JaversSpringDataAuditable
 @Repository
-public interface LocationRepository extends CachingJpaRepository<Location, Integer>,
-                                            SnapshotRepository<Location, Integer> {
+public interface LocationRepository extends CachingJpaRepository<Location, Long>,
+                                            SnapshotRepository<Location, Long> {
 }
